Implement a new tab that collects all the Portfolio Images and shows them, allowing the user to select one of them.
This is likely to be inappropriate when we have thousands of images: All
the images are read from the server when the GUi starts.  At 151 images
(today), it takes the better part of a minute to complete.  It may be
better to launch this as a separate application, when needed.