Two big changes.
1. Enable a new "Docents" tab, to satisfy the requirements of the
Docents for a simpler way to select their favorite channels and
pictures.

2. (Finally!) Implement the ability for the server and for the Display
itself to send feedback on what just happened.  in particular, the
Server now replies with an ERROR-type message if the selector does not
have the right to change a specific display or if its signature is
messed up.