Remove the check for a String being sent to the server, bundle that in with the other unknown objects (a String has not been transmitted received in 6 months).
If a client tries to connect and it has the same name as a different
client, remove BOTH of them from the list.  The experience on this has
been that this situation occurs only (a) when the user tries to launch a
second ChannelSelector (and now the onus is on her to fix this
situation), or (b) in a bug situation when a client tries to disconnect
and the server fails to recognize it, and this this same client (in a
different process after, say, a reboot) tries to connect again.  This
second situation should be fixed with this change.