This is a pretty big change.  The motivation is that it is not possible to detect when a client on a socket is killed due to, say, a sudden reboot. This code seems to handle all the other situations, but not the reboot one. (I tested "kill -9" on a client, and that seemed to be caught by this code.) See, for example, http://stackoverflow.com/questions/969866/java-detect-lost-connection.
The idea is that the server will keep track of which clients have been
pinged and have not responded.  It will be rather harsh in saying
goodbye to unresponsive clients!  Specifically, it will take no less
than 12 seconds to reject a tardy client, and no more than about (12
+ 1.5*(no. clients)) seconds.

So, the specific changes are:

* Add a shutdown hook on all messaging client objects

* Add a "No. of unsuccessful pings" counter to all connected clients. 
The server will FORGET these clients when this number is bigger than 4,
and it will reset this counter when the client responds.

* Move and cleanup calls to setLastSeen()

* Add a plethora of debug print statements.  I'll leave these in for a
while, to be sure that things are working well.  They will be removed in
a later release.

* Shorten the "too old" value for clients

* when ping'ing the oldest non-responsive client, do not ping the same
client twice in a row.